# functions/swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Bet Chicago
host: localhost:5001
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
  - application/octet-stream
  - application/vnd.contentful.management.v1+json
  - application/x-www-form-urlencoded
  - multipart/form-data
  - text/plain; charset=utf-8
  - image/jpeg
produces:
  - application/json
paths:
...
  /mlb/playerdata:
    x-swagger-router-controller: mlb
    post:
      description: update firedata from sports radar
      operationId: mlbPlayerData
      tags:
        - MLB
      parameters:
        - name: playerId
          in: query
          description: SR Player Id - 919d18a0-cd70-43c9-81f2-b96bf9e66a17
          required: true
          type: string
        - name: type
          in: query
          description: profile or pitching
          required: true
          type: string
          enum:
            - profile
            - pitching
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /mlb/teamdata:
    x-swagger-router-controller: mlb
    post:
      description: update firedata from sports radar
      operationId: mlbTeamData
      tags:
        - MLB
      parameters:
        - name: teamId
          in: query
          description: SR Team Id - 03556285-bdbb-4576-a06d-42f71f46ddc5
          required: true
          type: string
        - name: type
          in: query
          description: profile, depth_chart
          required: true
          type: string
          enum:
            - profile
            - depth_chart
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /mlb/teamseasondata:
    x-swagger-router-controller: mlb
    post:
      description: update firedata from sports radar
      operationId: mlbTeamSeasonData
      tags:
        - MLB
      parameters:
        - name: teamId
          in: query
          description: SR Team Id - 03556285-bdbb-4576-a06d-42f71f46ddc5
          required: true
          type: string
        - name: type
          in: query
          description: splits, statistics
          required: true
          type: string
          enum:
            - splits
            - statistics
        - name: year
          in: query
          description: year 2018, 2017, ... 2012 - maybe farther
          required: true
          type: string
        - name: season
          in: query
          description: pre, regular, post
          required: true
          type: string
          enum:
            - REG
            - PRE
            - PST
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /mlb/dailydata:
    x-swagger-router-controller: mlb
    post:
      description: update firedata from sports radar
      operationId: updateMLBDailyData
      tags:
        - MLB
      parameters:
        - name: date
          in: query
          description: date if not current
          required: false
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /mlb/dailyboxscore:
    x-swagger-router-controller: mlb
    post:
      description: update firedata from sports radar for boxscore only
      operationId: updateMLBDailyBoxscore
      tags:
        - MLB
      parameters:
        - name: date
          in: query
          description: date if not current
          required: false
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /mlb/gamedata:
    x-swagger-router-controller: mlb
    post:
      description: update firedata from sports radar
      operationId: updateMLBGameData
      tags:
        - MLB
      parameters:
        - name: eventId
          in: query
          description: eventId
          required: true
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /mlb/yearlydata:
    x-swagger-router-controller: mlb
    post:
      description: update firedata from sports radar
      operationId: updateMLBYearData
      tags:
        - MLB
      parameters:
        - name: year
          in: query
          description: year
          required: true
          type: string
        - name: tag
          in: query
          description: yearly stats
          required: true
          type: string
          enum:
            - statistics
            - rankings
            - schedule
            - standings
            - transactions
            - injuries
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /mlb/scorepush:
    x-swagger-router-controller: mlb
    get:
      description: for long polling on sport radar mlb line score push feed
      operationId: handleBaseballPush
      tags:
        - MLB
      parameters:
        - name: type
          in: query
          description: stat type to feed
          required: true
          type: string
          enum:
            - statistics
            - linescore
            - events
      responses:
        "200":
          description: job completion status
          schema:
            type: string
...
  /nba/updateDailyData:
    x-swagger-router-controller: nba
    post:
      description: daily schedule data
      operationId: updateDailyData
      tags:
        - NBA
      parameters:
        - name: date
          in: query
          description: Date
          required: false
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /nba/updateSeasonStats:
    x-swagger-router-controller: nba
    post:
      description: Season Statistics
      operationId: updateSeasonStats
      tags:
        - NBA
      parameters:
        - name: year
          in: query
          description: Year
          required: true
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /nba/updateYearData:
    x-swagger-router-controller: nba
    post:
      description: Year schedule and standings data
      operationId: updateYearData
      tags:
        - NBA
      parameters:
        - name: year
          in: query
          description: Year
          required: true
          type: string
        - name: tag
          in: query
          description: schedule, standings
          required: true
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /nba/updateGameData:
    x-swagger-router-controller: nba
    post:
      description: Game data
      operationId: updateGameData
      tags:
        - NBA
      parameters:
        - name: gameId
          in: query
          description: Game Id
          required: true
          type: string
        - name: tag
          in: query
          description: boxscore, summary, pbp
          required: true
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /nba/updatePlayerData:
    x-swagger-router-controller: nba
    post:
      description: Player data
      operationId: updatePlayerData
      tags:
        - NBA
      parameters:
        - name: playerId
          in: query
          description: Player Id
          required: true
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /nba/updateTeamData:
    x-swagger-router-controller: nba
    post:
      description: Team data
      operationId: updateTeamData
      tags:
        - NBA
      parameters:
        - name: teamId
          in: query
          description: Team Id
          required: true
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /nba/getFutureData:
    x-swagger-router-controller: nba
    post:
      description: Ranking data
      operationId: getFutureData
      tags:
        - NBA
      parameters:
        - name: year
          in: query
          description: Year
          required: true
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
...